requirements.txt
asgiref==3.8.1
autopep8==2.3.2
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==5.1.7
django-crispy-forms==2.3
django-elasticsearch-dsl==8.0
django-filter==25.1
django-htmx==1.23.0
django-js-asset==3.1.2
django-mptt==0.16.0
django-silk==5.3.2
django-template-partials==24.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
elastic-transport==8.17.1
elasticsearch==8.17.2
elasticsearch-dsl==8.17.1
factory_boy==3.3.3
Faker==37.1.0
gprof2dot==2024.6.6
idna==3.10
iniconfig==2.1.0
Markdown==3.7
packaging==25.0
pillow==11.1.0
pluggy==1.5.0
pycodestyle==2.13.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
webencodings==0.5.1
